requirements-lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt -o requirements-lock.txt
anyio==4.13.0
    # via
    #   httpx
    #   jupyter-server
argon2-cffi==25.1.0
    # via jupyter-server
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
arrow==1.4.0
    # via isoduration
asttokens==3.0.1
    # via stack-data
async-lru==2.3.0
    # via jupyterlab
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.18.0
    # via jupyterlab-server
beautifulsoup4==4.14.3
    # via nbconvert
bleach==6.3.0
    # via nbconvert
cachecontrol==0.14.4
    # via itscalledsoccer
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.7
    # via requests
comm==0.2.3
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.20
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
executing==2.2.1
    # via stack-data
fastjsonschema==2.21.2
    # via nbformat
fqdn==1.5.1
    # via jsonschema
geographiclib==2.1
    # via geopy
geopy==2.4.1
    # via -r requirements.txt
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via jupyterlab
idna==3.11
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
ipykernel==7.2.0
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==9.12.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.8
    # via jupyter
isoduration==20.11.0
    # via jsonschema
itscalledsoccer==1.3.3
    # via -r requirements.txt
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.14.0
    # via jupyterlab-server
jsonpointer==3.1.1
    # via jsonschema
jsonschema==4.26.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyter==1.1.1
    # via -r requirements.txt
jupyter-client==8.8.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.9.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.12.0
    # via jupyter-server
jupyter-lsp==2.3.1
    # via jupyterlab
jupyter-server==2.17.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.4
    # via jupyter-server
jupyterlab==4.5.6
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.28.0
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.16
    # via ipywidgets
lark==1.3.1
    # via rfc3987-syntax
markupsafe==3.0.3
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.2.1
    # via
    #   ipykernel
    #   ipython
mistune==3.2.0
    # via nbconvert
msgpack==1.1.2
    # via cachecontrol
nbclient==0.10.4
    # via nbconvert
nbconvert==7.17.1
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.5.5
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.4.4
    # via
    #   -r requirements.txt
    #   pandas
    #   pgeocode
packaging==26.0
    # via
    #   ipykernel
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandas==3.0.2
    # via
    #   -r requirements.txt
    #   itscalledsoccer
    #   pgeocode
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.6
    # via jedi
pexpect==4.9.0
    # via ipython
pgeocode==0.5.0
    # via -r requirements.txt
platformdirs==4.9.6
    # via jupyter-core
prometheus-client==0.25.0
    # via jupyter-server
prompt-toolkit==3.0.52
    # via
    #   ipython
    #   jupyter-console
psutil==7.2.2
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pyarrow==23.0.1
    # via -r requirements.txt
pycparser==3.0
    # via cffi
pygments==2.20.0
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   jupyter-console
    #   nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   pandas
python-json-logger==4.1.0
    # via jupyter-events
pyyaml==6.0.3
    # via jupyter-events
pyzmq==27.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
rapidfuzz==3.14.5
    # via itscalledsoccer
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.33.1
    # via
    #   -r requirements.txt
    #   cachecontrol
    #   itscalledsoccer
    #   jupyterlab-server
    #   pgeocode
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rfc3987-syntax==1.1.0
    # via jsonschema
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
send2trash==2.1.0
    # via jupyter-server
setuptools==82.0.1
    # via jupyterlab
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
soupsieve==2.8.3
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.4.0
    # via bleach
tornado==6.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.15.0
    # via beautifulsoup4
tzdata==2026.1
    # via arrow
uri-template==1.3.0
    # via jsonschema
urllib3==2.6.3
    # via requests
wcwidth==0.6.0
    # via prompt-toolkit
webcolors==25.10.0
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.9.0
    # via jupyter-server
widgetsnbextension==4.0.15
    # via ipywidgets
